Switch 'rmap-api' and 'rmap-webapp' modules to using Postgres for tests. - Add docker-maven-plugin boilerplate - Update build-helper and failsafe plugin configurations - Rename unit tests so they are recognized as integration tests (and so they can fit into the Maven lifecycle for integration tests)
