+++ api/src/test/java/info/rmapproject/api/responsemgr/DiscoResponseManagerTestIT.java
@@ -45,7 +45,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import info.rmapproject.api.ApiDataCreationTestAbstract;
+import info.rmapproject.api.ApiDataCreationTestAbstractIT;
 import info.rmapproject.api.auth.ApiUserService;
 import info.rmapproject.api.auth.ApiUserServiceImpl;
 import info.rmapproject.api.exception.ErrorCode;
@@ -71,7 +71,7 @@ import info.rmapproject.testdata.service.TestFile;
  * Tests for DiscoResponseManager
  * @author khanson
  */
-public class DiscoResponseManagerTest extends ApiDataCreationTestAbstract {
+public class DiscoResponseManagerTestIT extends ApiDataCreationTestAbstractIT {
 			
 	@Autowired
 	protected Timegate timegate;
