+++ webapp/pom.xml
@@ -114,8 +114,8 @@
 		    <scope>test</scope>
 		</dependency>
         <dependency>
-            <groupId>org.apache.derby</groupId>
-            <artifactId>derby</artifactId>
+            <artifactId>postgresql</artifactId>
+            <groupId>org.postgresql</groupId>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -155,11 +155,32 @@
                 <artifactId>maven-surefire-plugin</artifactId>
                 <configuration>
                     <systemProperties>
-                        <derby.system.home>${project.build.testOutputDirectory}/derby</derby.system.home>
                         <spring.profiles.active>default,inmemory-db,inmemory-idservice,inmemory-triplestore,embedded-solr</spring.profiles.active>
                         <logback.configurationFile>src/test/resources/logback-test.xml</logback.configurationFile>
+                        <docker.host.address>${docker.host.address}</docker.host.address>
+                        <db.port>${db.port}</db.port>
                     </systemProperties>
                 </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-failsafe-plugin</artifactId>
+                <configuration>
+                    <systemProperties>
+                        <spring.profiles.active>default,inmemory-db,inmemory-idservice,inmemory-triplestore,embedded-solr</spring.profiles.active>
+                        <logback.configurationFile>src/test/resources/logback-test.xml</logback.configurationFile>
+                        <docker.host.address>${docker.host.address}</docker.host.address>
+                        <db.port>${db.port}</db.port>
+                    </systemProperties>
+                </configuration>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>integration-test</goal>
+                            <goal>verify</goal>
+                        </goals>
+                    </execution>
+                </executions>
             </plugin>
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
@@ -180,6 +201,55 @@
                                 </resource>
                             </resources>
                         </configuration>
+                    </execution>
+                    <execution>
+                        <id>reserve-port</id>
+                        <phase>validate</phase>
+                        <goals>
+                            <goal>reserve-network-port</goal>
+                        </goals>
+                        <configuration>
+                            <portNames>
+                                <portName>db.port</portName>
+                            </portNames>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <groupId>io.fabric8</groupId>
+                <artifactId>docker-maven-plugin</artifactId>
+                <configuration>
+                    <images>
+                        <image>
+                            <alias>postgres</alias>
+                            <name>postgres:latest</name>
+                            <run>
+                                <cmd>postgres</cmd>
+                                <ports>
+                                    <port>db.port:5432</port>
+                                </ports>
+                                <wait>
+                                    <log>ready to accept connections|database system is shut down</log>
+                                </wait>
+                            </run>
+                        </image>
+                    </images>
+                </configuration>
+                <executions>
+                    <execution>
+                        <id>start-docker-for-tests</id>
+                        <phase>pre-integration-test</phase>
+                        <goals>
+                            <goal>start</goal>
+                        </goals>
+                    </execution>
+                    <execution>
+                        <id>stop-docker-after-tests</id>
+                        <phase>post-integration-test</phase>
+                        <goals>
+                            <goal>stop</goal>
+                        </goals>
                     </execution>
                 </executions>
             </plugin>
