+++ webapp/src/test/java/info/rmapproject/webapp/controllers/HomeControllerTestIT.java
@@ -1,0 +1,77 @@
+/*******************************************************************************
+ * Copyright 2017 Johns Hopkins University
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * This software was produced as part of the RMap Project (http://rmap-project.info),
+ * The RMap Project was funded by the Alfred P. Sloan Foundation and is a 
+ * collaboration between Data Conservancy, Portico, and IEEE.
+ *******************************************************************************/
+package info.rmapproject.webapp.controllers;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+import info.rmapproject.webapp.WebTestAbstractIT;
+import info.rmapproject.webapp.utils.Constants;
+
+/**
+ * Basic tests for HomeController, which returns views for home and contact page.
+ * @author khanson
+ *
+ */
+public class HomeControllerTestIT extends WebTestAbstractIT {
+	
+    @Autowired
+    private WebApplicationContext wac;
+    
+    private MockMvc mockMvc;
+
+    @Before
+    public void init() {
+        mockMvc = MockMvcBuilders.webAppContextSetup(wac).build();
+    }
+
+    /**
+     * Quick test for home page
+     * @throws Exception
+     */
+    @Test
+    public void testGetHomePage() throws Exception {
+        mockMvc.perform(get("/home"))
+                .andExpect(status().isOk())
+                .andExpect(view().name("home"))
+                .andExpect(model().attributeExists(Constants.SITE_PROPERTIES_ATTRIBNAME));        
+    }
+
+    /**
+     * Quick test for home page
+     * @throws Exception
+     */
+    @Test
+    public void testGetAboutPage() throws Exception {
+        mockMvc.perform(get("/about"))
+                .andExpect(status().isOk())
+                .andExpect(view().name("about"))
+                .andExpect(model().attributeExists(Constants.SITE_PROPERTIES_ATTRIBNAME));      
+    }
+}
