--- webapp/src/test/resources/derby/derby.properties
@@ -1,1 +1,0 @@
-derby.stream.error.logSeverityLevel=0
